FLOAT is float, FLOAT8 is double  again ;-)

+++ libmp3lame/machine.h
@@ -88,7 +88,7 @@
 #	define WIN32_LEAN_AND_MEAN
 #	include <windows.h>
 #else
-	typedef double FLOAT;
+	typedef float FLOAT;
 #endif
 
 /*
@@ -103,7 +103,7 @@ lot of conversions.
 */
 
 /* NOTE: RH: 7/00:  if FLOAT8=float, it breaks resampling and VBR code */
-typedef float FLOAT8;
+typedef double FLOAT8;
